split off operations to separate module
