+++ src/main.rs
@@ -23,16 +23,7 @@
  *
  */
 
-use std::error::Error;
-use std::fs::File;
-use std::sync::mpsc;
-use std::thread::sleep;
-use std::time::Duration;
-
 mod error;
-use crate::error::*;
-
-use log::{debug, error, info, warn};
 
 #[macro_use]
 mod util;
@@ -40,114 +31,7 @@ mod util;
 mod mpkbank;
 mod mpkmidi;
 mod u14;
-
-use crate::mpkbank::MpkBankDescriptor;
-use crate::mpkmidi::*;
-use crate::util::*;
-
-fn snoop() -> Result<(), Box<dyn Error>> {
-    let cb = |_, bytes: &[u8], _: &mut _| {
-        debug!("rx bytes: {:?}", bytes);
-        match MpkMidiMessage::parse_msg(bytes) {
-            Ok(m) => println!("{:?}", m),
-            Err(e) => warn!("Unparsed: {}; bytes: {:?}", e, bytes),
-        }
-    };
-    let _midi_in = midi_in_connect(cb, ())?;
-    info!("Snoop started. Use CTRL-C to stop.");
-    loop {
-        sleep(Duration::from_millis(250));
-    }
-}
-
-fn passthrough() -> Result<(), Box<dyn Error>> {
-    let (tx, rx) = mpsc::channel();
-
-    let cb = move |_, bytes: &[u8], _: &mut _| {
-        debug!("rx bytes: {:?}", bytes);
-        match MpkMidiMessage::parse_msg(bytes) {
-            Ok(m) => println!("{:?}", m),
-            Err(e) => warn!("Unparsed: {}; bytes: {:?}", e, bytes),
-        }
-        if let Err(e) = tx.send(Vec::from(bytes)) {
-            error!("Error while sending: {}", e);
-        }
-    };
-
-    let mut midi_out = midi_out_connect()?;
-    let _midi_in = midi_in_connect(cb, ())?;
-
-    info!("Passthrough started: MIDI messages from input will be sent to output. Use CTRL-C to stop.");
-    loop {
-        match rx.recv() {
-            Ok(m) => {
-                if let Err(e) = midi_out.send(m.as_slice()) {
-                    error!("Error while forwarding: {}", e);
-                }
-            }
-            Err(e) => {
-                error!("Error while receiving: {}", e);
-            }
-        }
-    }
-}
-
-fn get_bank_desc(bank: u8) -> Result<MpkBankDescriptor, Box<dyn Error>> {
-    check_bank_value!(bank);
-
-    let (tx, rx) = mpsc::channel();
-
-    let cb = move |_, bytes: &[u8], _: &mut _| {
-        if let Ok(m) = MpkMidiMessage::parse_msg(bytes) {
-            if let MpkMidiMessage::Bank(bank_rx, d) = m {
-                if bank != bank_rx {
-                    error!("Error: received bank {}, expected {}", bank_rx, bank);
-                }
-                if let Err(e) = tx.send(d) {
-                    error!("Error while sending on channel: {}", e);
-                }
-            } else {
-                warn!("Unexpected message (ignored): {:?}", m);
-            }
-        } else {
-            warn!("Unparsed: {:?}", bytes);
-        }
-    };
-
-    let mut midi_out = midi_out_connect()?;
-    let midi_in = midi_in_connect(cb, ())?;
-
-    midi_out.send(sysex_get_bank(bank).as_slice())?;
-    let bank_desc = rx.recv_timeout(Duration::new(10, 0))?;
-
-    midi_out.close();
-    midi_in.close();
-
-    Ok(bank_desc)
-}
-
-fn set_bank_from_desc(bank: u8, bank_desc: MpkBankDescriptor) -> Result<(), Box<dyn Error>> {
-    check_bank_value!(bank);
-
-    let mut midi_out = midi_out_connect()?;
-    midi_out.send(&sysex_set_bank(bank, bank_desc))?;
-    midi_out.close();
-
-    Ok(())
-}
-
-fn show_bank(bank: u8) -> Result<(), Box<dyn Error>> {
-    let bank_desc = get_bank_desc(bank)?;
-    println!("Bank {}:\n{}", bank, bank_desc);
-    Ok(())
-}
-
-fn dump_bank_yaml(bank: u8) -> Result<(), Box<dyn Error>> {
-    let bank_desc = get_bank_desc(bank)?;
-    let serialized = serde_yaml::to_string(&bank_desc).unwrap();
-    println!("{}", serialized);
-    Ok(())
-}
+mod operations;
 
 // fn cmd_show_bank(matches: &ArgMatches) -> Result<(), Box<dyn Error>> {
 //     let bank = matches.value_of("bank").unwrap().parse::<u8>()?;
